remove libevhtp
