+++ qa/common/check_copyright.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# Copyright (c) 2018-2019, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2018-2020, NVIDIA CORPORATION. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -36,7 +36,6 @@ SKIP_EXTS = ('jpeg', 'jpg', 'pgm', 'png',
              'preprocessed', 'jmx', 'gz',
              'caffemodel', 'json')
 SKIP_PATHS = ('builddir',
-              'build/libevhtp',
               'build/onnxruntime',
               'deploy/single_server/.helmignore',
               'docs/examples/model_repository',
