+++ src/servers/CMakeLists.txt
@@ -73,42 +73,28 @@ set(HTTP_ENDPOINT_OBJECTS "")
 set(HTTP_ENDPOINT_LIBRARIES "")
 
 if(${TRTIS_ENABLE_HTTP} OR ${TRTIS_ENABLE_METRICS})
-  if(NOT ${TRTIS_ENABLE_HTTP_H2O})
-    find_package(libevhtp CONFIG REQUIRED)
-    message(STATUS "Using libevhtp ${libevhtp_VERSION}")
-
-    if(${TRTIS_ENABLE_HTTP_V2})
-      set(
-        HTTP_ENDPOINT_SRCS
-        http_v2_server.cc
-      )
-    else()
-      set(
-        HTTP_ENDPOINT_SRCS
-        http_server.cc
-      )
-    endif() # TRTIS_ENABLE_HTTP_V2
-
-    set(
-      HTTP_ENDPOINT_HDRS
-      http_server.h
-    )
-  else()
-    find_library(H2O_LIB h2o REQUIRED)
-    message(STATUS "Using libh2o ${H2O_LIB}")
-    find_library(UV_LIB uv REQUIRED)
-    message(STATUS "Using libuv ${UV_LIB}")
-    find_package(Threads REQUIRED)
-    find_package(OpenSSL REQUIRED)
+  find_library(H2O_LIB h2o REQUIRED)
+  message(STATUS "Using libh2o ${H2O_LIB}")
+  find_library(UV_LIB uv REQUIRED)
+  message(STATUS "Using libuv ${UV_LIB}")
+  find_package(Threads REQUIRED)
+  find_package(OpenSSL REQUIRED)
+  if(${TRTIS_ENABLE_HTTP_V2})
     set(
       HTTP_ENDPOINT_SRCS
-      http_h2o_server.cc
+      http_v2_server.cc
     )
+  else()
     set(
-      HTTP_ENDPOINT_HDRS
-      http_server.h
+      HTTP_ENDPOINT_SRCS
+      http_server.cc
     )
-  endif() # TRTIS_ENABLE_HTTP_H2O
+  endif() # TRTIS_ENABLE_HTTP_V2
+
+  set(
+    HTTP_ENDPOINT_HDRS
+    http_server.h
+  )
 
   add_library(
     http-endpoint-library EXCLUDE_FROM_ALL OBJECT
@@ -121,18 +107,11 @@ if(${TRTIS_ENABLE_HTTP} OR ${TRTIS_ENABLE_METRICS})
 
   add_dependencies(http-endpoint-library proto-library)
 
-  if(NOT ${TRTIS_ENABLE_HTTP_H2O})
-    target_include_directories(
-      http-endpoint-library
-      PRIVATE $<TARGET_PROPERTY:libevhtp::evhtp,INTERFACE_INCLUDE_DIRECTORIES>
-    )
-  else()
-    target_include_directories(
-      http-endpoint-library
-      PRIVATE -lh2o
-      PRIVATE -luv
-    )
-  endif() # TRTIS_ENABLE_HTTP_H2O
+  target_include_directories(
+    http-endpoint-library
+    PRIVATE -lh2o
+    PRIVATE -luv
+  )
 
   set(
     HTTP_ENDPOINT_OBJECTS
@@ -140,41 +119,31 @@ if(${TRTIS_ENABLE_HTTP} OR ${TRTIS_ENABLE_METRICS})
     $<TARGET_OBJECTS:proto-library>
   )
 
-  if(NOT ${TRTIS_ENABLE_HTTP_H2O})
-    if(${TRTIS_ENABLE_HTTP_V2})
-      set(
-        HTTP_ENDPOINT_OBJECTS
-        ${HTTP_ENDPOINT_OBJECTS}
-        $<TARGET_OBJECTS:grpc-library>
-      )
-    endif() # TRTIS_ENABLE_HTTP_V2
+  if(${TRTIS_ENABLE_HTTP_V2})
+    set(
+      HTTP_ENDPOINT_OBJECTS
+      ${HTTP_ENDPOINT_OBJECTS}
+      $<TARGET_OBJECTS:grpc-library>
+    )
+  endif() # TRTIS_ENABLE_HTTP_V2
+  set(
+    HTTP_ENDPOINT_LIBRARIES
+    PRIVATE h2o
+    PRIVATE h2o-evloop
+    PRIVATE uv
+    PRIVATE -lwslay
+    PRIVATE protobuf::libprotobuf
+    PRIVATE -lre2
+    PRIVATE -lnuma
+  )
+  if(${TRTIS_ENABLE_HTTP_V2})
     set(
       HTTP_ENDPOINT_LIBRARIES
-      PRIVATE ${LIBEVENT_LIBRARIES}
-      PRIVATE libevhtp::evhtp
-      PRIVATE protobuf::libprotobuf
-      PRIVATE -lre2
+      ${HTTP_ENDPOINT_LIBRARIES}
+      PRIVATE gRPC::grpc++
+      PRIVATE gRPC::grpc
     )
-    if(${TRTIS_ENABLE_HTTP_V2})
-      set(
-        HTTP_ENDPOINT_LIBRARIES
-        ${HTTP_ENDPOINT_LIBRARIES}
-        PRIVATE gRPC::grpc++
-        PRIVATE gRPC::grpc
-      )
-    endif() # TRTIS_ENABLE_HTTP_V2
-  else()
-    set(
-      HTTP_ENDPOINT_LIBRARIES
-      PRIVATE h2o
-      PRIVATE h2o-evloop
-      PRIVATE uv
-      PRIVATE -lwslay
-      PRIVATE protobuf::libprotobuf
-      PRIVATE -lre2
-      PRIVATE -lnuma
-    )
-  endif() # TRTIS_ENABLE_HTTP_H2O
+  endif() # TRTIS_ENABLE_HTTP_V2
 endif() # TRTIS_ENABLE_HTTP || TRTIS_ENABLE_METRICS
 
 #
